Renamed roleLog to logging.SerialLogger
Fixed compilation after file moved.
